Twitter.com (Pals: Fiora Aeterna, Tom Forsyth, Fabian Giesen, Matthew VanDevander, Tom Francis, Allen Webster, Chronaldragon, Sean Barret, AbnerCoimbre and others contribution!!) ...contribution!! Handmade Hero (Streamer: Casey Muratori...) ... contribution)!! Livecoding.tv contribution (Streamers: Vaapamies, ThomasBS, epryka, chocsx, Pals: contribution: HARDTriper, JeffersonCiotti,jbvsl, r00tKiller, and many others learning, watching, cheering up or staging contribution...) ... contribution!! from hyperv (multiboot)!! Windows 10 contribution!! FreeRDP contribution!! Search Oracles contribution!! Simone e Fábio contribution!! Carol contributin!!
From X:\core at inha hyperv at veriverse!!
｛
｝

   ----
                        -   -   -
   Temp(TODO annotation) : Removing statically initialized variables. Leaving only "integer" types
                           static globals and "integer"->pointer types static globals with Sys(::get_thread_app())
						   balm (please register at Sys[System] the initialized objects  so it can be deleted
						   fashionly (you can take care of order,before alloc system is deleted and even between
						   single tons, when dependency information is registered also (great IDEA : great TODO!!)
Mummi contribution!!, eu te amo demais!!
-----8nnnnnnnn8
++++++++++

+++ appseed/draw2d_gdiplus/draw2d_gdiplus_path.cpp
@@ -350,7 +350,7 @@ namespace draw2d_gdiplus
       format.SetLineAlignment(Gdiplus::StringAlignmentNear);
 
 
-      m_ppath->AddString(wstr,wstr.get_length(),&fontFamily,iStyle,dSize,Gdiplus::Point(x,y),&format);
+      m_ppath->AddString(wstr, (INT) wstr.get_length(),&fontFamily,iStyle,dSize,Gdiplus::Point(x,y),&format);
 
       return true;
